+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Version 0.7.2 (TBD)
 * Added more detailed information to the server and plugin log files about plugin errors.
+* Fixed a bug where `TrackingMultimap#percentKeyDataType` returned `NaN` instead of `0` when the map was empty.
 
 #### Version 0.7.1 (November 22, 2017)
 * Fixed a bug that caused an error in some cases of importing or inserting data that contained a value of `-`.